requirements.txt
requests
types-requests
mypy
pymupdf-layout
pymupdf4llm
tqdm
types-tqdm
